#7480 : apply mdoubez's patch for lib2geom compilation under msvc
git-svn-id: svn://scribus.net/trunk/Scribus@12665 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/plugins/tools/2geomtools/meshdistortion/meshdistortiondialog.cpp
@@ -315,7 +315,7 @@ MeshDistortionDialog::MeshDistortionDialog(QWidget* parent, ScribusDoc *doc) : Q
 						if(vi == 0 && ui == 0)
 							base = Geom::Point(w4, w4);
 						double dl = dot((handles[corner+4*i] - base), dir)/dot(dir,dir);
-						sb2[dim][i][corner] = dl/(ww/2)*pow(4.0,ui+vi);
+						sb2[dim][i][corner] = dl/(ww/2)*pow(4.0f,((int)(ui+vi)));
 					}
 				}
 			}
@@ -431,7 +431,7 @@ void MeshDistortionDialog::updateMesh(bool gridOnly)
 						if(vi == 0 && ui == 0)
 							base = Geom::Point(w4, w4);
 						double dl = dot((handles[corner+4*i] - base), dir)/dot(dir,dir);
-						sb2[dim][i][corner] = dl/(ww/2)*pow(4.0,ui+vi);
+						sb2[dim][i][corner] = dl/(ww/2)*pow(4.0f,((int)(ui+vi)));
 					}
 				}
 			}
